requirements.txt
anyio==3.6.2
click==8.1.3
fastapi
h11==0.14.0
httptools==0.5.0
idna==3.4
pydantic
python-dotenv==0.21.1
PyYAML==6.0
sniffio==1.3.0
starlette==0.19.1
typing_extensions
uvicorn
watchfiles==0.18.1
websockets==10.4
langchain
openai
pandas
tabulate